Update lotus 1.18 (#163)
* Update lotus 1.18 rc3

* circle ci fix

* Fix

* rollback

* Action

* Github action

* Fixed type

* Go sum

* Github action

* Github action

* Fixed some tests

* Actions

* update go state types

* Go states

* Fix action

* Actions

* Update to rc4

* Added supported operations into networkapiservice struct

* Update to rc5

* Added supported operations into networkapiservice struct

* Modified rosetta-cli config

* Modified rosetta-cli config

* Update rosetta-cli

* Modified config

* Modified config

* Modified config

* Bumped rosetta-filecoin-lib

+++ rosetta/services/construction.go
@@ -168,7 +168,7 @@ func (c *ConstructionAPIService) ConstructionMetadata(
 
 			// GasEstimateMessageGas to get a safely overestimated value for gas limit
 			message, err = c.node.GasEstimateMessageGas(ctx, message,
-				&api.MessageSendSpec{MaxFee: filTypes.NewInt(build.BlockGasLimit)}, filTypes.TipSetKey{})
+				&api.MessageSendSpec{MaxFee: filTypes.NewInt(uint64(build.BlockGasLimit))}, filTypes.TipSetKey{})
 			if err != nil {
 				return nil, BuildError(ErrUnableToEstimateGasLimit, err, true)
 			}
